ssa_builder perf
Summary:
The SSA builder had a serious perf issue on large files. The common operation of
merging two SSA values would create a SSA value that might keep multiple copies
of the same subterm, eventually causing exponential blowup in the size of data
structures.

This diff implements a partial solution that seems to work well enough in
practice. Implementing a full solution shouldn't be too hard: it's documented in
comments.

Reviewed By: samwgoldman

Differential Revision: D5832387

fbshipit-source-id: 654e44f3992138cb0c7f816e82f50bd6d2d1bfca