Calculate exports, type exports, and declared modules before infer
Summary:
Calculating requires before infer made lazy type checking possible. This diff
adds a similar analysis that extracts exports.

The behavior is copied from the existing analysis in statement.ml, which
determines a file's module_kind (CJS or ES) and any declared modules.

This analysis is a little more granular, in that type exports are separate from
value exports.

This analysis could be used to break up large cycles, but my current plan is to
replace the analysis in statement.ml as a step toward removing local inference
entirely.

Reviewed By: mroch

Differential Revision: D5225066

fbshipit-source-id: ee5989c422a762de2eab4593f98f5e6acfb17772