Differentiate CJS require and ES import in file sigs
Summary:
In order to use file sigs to set up context dependencies, it's useful to
differentiate between CJS and ES import styles. In particular, we want to be
able to fire hooks (there are separate hooks for each style).

I will follow up with another change to add more structure on the ES import
side, so we can know precisely which parts of a dependency a module depends on.

Reviewed By: mroch

Differential Revision: D5835353

fbshipit-source-id: c76b831ad19cb8bf3da88d84be17d7d4eee54fae