Add $Rest types with the semantics of runtime object rest
Summary:
> Note: This commit only provides a `$Rest` type. It does not re-implement our support of object destructuring. However, the plan is to re-implement our object destructuring code in terms of `$Rest` around the same time we re-implement our spread code in terms of object type spread (`type T = {...A, ...B}`). The main win for this diff is that we get a better `$Diff` implementation in an upcoming commit.

`$Diff` is currently only usable in an upper bound position. `$Diff` also has no runtime equivalent which makes it unsound to use `$Diff` to model any runtime operation. These two factors together make `$Diff` very difficult to use.

This diff (pun intended) introduces `$Rest`. A utility type that soundly models runtime object rest behavior. For example:

```
const {a, b, ...rest} = (o: O);
(rest: $Rest<O, {|a: any, b: any|}>)
```

In a later diff I re-implement `$Diff` to use an unsound version of `$Rest`.

Like `$Diff`, `$Rest` takes two type arguments. Both of which should be objects. However, at runtime you list your properties and don't associate them with a type. So what is the intuition for the types of properties in the second object for `$Rest`? Take the following example where we check the types of `a` and `b` after destructuring them:

```
const {a, b, ...rest} = (o: O);
(a: A);
(b: B);
```

This is the same as:

```
$Rest<O, {|a: A, b: B|}>
```

`$Rest` will check the type of `a` against `A` and `b` against `B` as if you checked the types of `a` and `b` yourself after destructuring `O`.

Note that the second type argument is exact. When you have values (`const {a, b, ...rest}`) then the object `{|a, b|}` will always be exact. However, what is the intuition for an inexact second argument to `$Rest`? To answer this question we create a hypothetical implementation for `function rest<A, B>(a: A, b: B): $Rest<A, B>` as:

```
function rest<A, B>(a: A, b: B): $Rest<A, B> {
  const result = {};
  for (const k in a) {
    if (hasOwnProperty(a, k)) {
      if (\!hasOwnProperty(b, k)) {
        result[k] = a[k];
      }
    }
  }
  return result;
}
```

An inexact object may have any number of properties besides the one defined in its type. So the result of some `$Rest` usage: `$Rest<{|a: number, b: number|}, {}>` is: `{|a?: number, b?: number|}` because `{}` may have an `a` property or a `b` property.

Similarly, notice how we use `hasOwnProperty()`. If `a` has a property on its prototype chain then we will not add it to result. If `b` has a property on its prototype chain then we will not skip it if `a` has the same, own, property.

More documentation on this feature will come later. There are a few more object utility types I want to explore before writing documentation.

Reviewed By: samwgoldman

Differential Revision: D5801907

fbshipit-source-id: 339ad3d9fcec1b31761963f3d6e0d78e7f2b5687