Print full type error in Try Flow
Summary:
Before this change, we only printed the top-most error on Try Flow.
Increasingly, we are using the "extra" field of an error to provide essential
context for an error.

For example, given the code:

```
type A = { p: { q: string } };
type B = { +p: { +q: number } };
declare var a: A;
(a: B);
```

Try flow would only print:

```
4: (a: B);
    ^ object type. This type is incompatible with
4: (a: B);
       ^ object type
```

After this change, we see the full error:

```
4: (a: B);
    ^ object type. This type is incompatible with
4: (a: B);
       ^ object type
  Property `p` is incompatible:
    1: type A = { p: { q: string } };
                     ^ object type. This type is incompatible with
    2: type B = { +p: { +q: number } };
                      ^ object type
      Property `q` is incompatible:
        1: type A = { p: { q: string } };
                              ^ string. This type is incompatible with
        2: type B = { +p: { +q: number } };
                                ^ number
```

Reviewed By: mroch

Differential Revision: D5860352

fbshipit-source-id: bddce1a8eca450a5bb6e0eeaeba3e3659034b28f