suppress more useless ops
Summary:
pointing at "call of method foo", when the error would have pointed at
the return value of said call (the identical range), or to one of the args, is
not helpful. especially in the former case, it's confusing to complain about the
call, when it's really however the return value gets used that is the problem.

so, we suppress RMethodCall ops when they overlap with the location of the real
error. there are very likely other times we could suppress them but this
incrementally removes ops without degrading anything.

Reviewed By: samwgoldman

Differential Revision: D5861238

fbshipit-source-id: d09b3c3fd0a0d1cd8012b14e9311b188e64d1a94