Pin js_of_ocaml version
Looks like the newest version (>= 3.0) of `js_of_ocaml` changes one of
the APIs we're using. I had to manually pin this to version 2.8.1 to get
it to work.

It looks like we're also manually pinning this version in the Travis
build (`resources/travis/install_deps.sh`). I'm new to OCaml; is there a
better way to declare a projects `opam` dependencies than a note in the
README?