Add dropdown to choose Vim keybindings
We have to load a few more scripts on the page. In total, this adds 80
KB when loading the `/try/` page. I didn't spend the time figure out how
to make this load asynchronously on selecting "Vim" because I figured
this was all dwarfed by the time it takes to load our 2.8 MB flow.js
script.

Also, the default way to build the website didn't work for me right
away. Instead, I had to run these commands:

    rm -rf website/statics/master/{flow.js,flowlib}
    cp bin/flow.js website/static/master/flow.js
    cp -r lib website/static/master/flowlib
    cd website
    make

Steps to reproduce: I tried to build the site from a fresh clone of the
repo, and followed the commands listed in `website/README.md`.

I'm not sure if this is an issue we should look into more broadly.