clean up Span
Summary:
generalized `Loc.contains` into a compare function (`span_compare`),
and then used it to implement `SpanMap.compare`.

Since `Span.t` is just a `Loc.t`, and can't be made opaque because `Loc.t` is
used as a key to `SpanMap`, the other helper functions are redundant with their
`Loc` counterparts.

After all that, `Span` is rather empty :)

Reviewed By: gabelevi

Differential Revision: D5880567

fbshipit-source-id: addb42b125934e17fe1de37c3e17b1fadf2cba59