Add typings for ReactDOM.createPortal()
Summary:
Adds typings for the new React 16 portals feature. Portals introduce a new kind of React node. The React portal. We model this as an opaque type and add it as a variant to our `React.Node` type. The React portal type is exported as `React.Portal`.

Once React has documentation for portals we should add the `React.Portal` type to https://flow.org/en/docs/react/types with a link to the React documentation.

Reviewed By: gaearon

Differential Revision: D5815692

fbshipit-source-id: 72149ebc17cb2e40a493bd8a553c24a45215237c