Test flow ls on non-existent files & add --imaginary flag
Summary:
When you call `flow ls` you can pass in 0 or more files or
directories. Files that don't exist should just be ignored. {D5884160}
broke that behavior a little bit so this fixes it and add tests.

And then I'm going a little further and adding the `--imaginary` flag
(let me know if you have a better name). One of the main way tooling
will want to use `flow ls` is to ask "Hey Flow, does root X care about
files A, B, C?" This works fine if A, B, and C are on disk. But maybe
you have some CI which is checking a diff which deletes A, B, and C. So
instead we're asking speculatively "Would flow care about A, B, and C if
they existed on disk".

Reviewed By: samwgoldman

Differential Revision: D5884535

fbshipit-source-id: 59f8486e34d73d8af3cecafc49a23d0c92567064