Make non-lazy mode check node_modules/ lazily
Summary:
calebmer pointed out that there's no real reason for Flow to check files in `node_modules/` which are not direct or transitive dependencies of the actual code we want to check. He thought we could perhaps reuse some of our lazy infrastructure to avoid checking this code. This is my attempt at that.

The basic idea is this:

* Previously non-lazy mode would treat every file in the repo and every file in node_modules as focused
* After this change, non-lazy mode treats every file in the repo as focused. Files in node_modules are ignored unless they are direct or transitive dependencies of files in the repo, in which case they are treated as dependencies.

I'm assuming that files in `node_modules/` can never be dependents of the focused files. This saves a ton of perf, since looking up dependents is really expensive. I hope there isn't some weird linking thing I need to worry about.

It seems pretty perf neutral. However it's probably worth making this change to avoid complaining about Flow errors in `node_modules/` code which is never imported.

Reviewed By: samwgoldman

Differential Revision: D5863553

fbshipit-source-id: b1fefba389d5b5cfb95830844f25469282f384c7