Refactor type destructors to support tvars
Summary:
Currently, Flow doesn't have great support for tvars in type destructors. This leads to issues when using type destructors like `$PropertyType`, `$Diff`, and `$ObjMap` with polymorphic function type variables which could easily be tvars. For example, the following does not produce an error when it should:

```
declare function fn<T>(x: (T) => void): ($NonMaybeType<T>) => void;

(fn((x: number) => {}): (string) => void);
```

If you remove the `$NonMaybeType<T>` then you correctly get an error.

This diff refactors type destructors such that we can support tvars.

We do this by introducing `TypeDestructorTriggerT (destructor, tout)`. We put a `TypeDestructorTriggerT (destructor, tout)` on both sides of a type so that we can see all new upper and lower bounds that are added to the type. So in our case above we would emit the constraints:

```
T ~> TypeDestructorTriggerT (NonMaybe, tout)
TypeDestructorTriggerT (NonMaybe, tout) ~> T
```

When `T` gets an upper bound of `number` our lower bound `TypeDestructorTriggerT` constraint would fire. This differs from the current implementation because we did not look at upper bounds of a tvar at all.

When our `TypeDestructorTriggerT` is triggered then we destruct `number` with `NonMaybe` and add it as an upper bound to `tout`.

```
TypeDestructorTriggerT (NonMaybe, tout) ~> number
  tout ~> NonMaybe(number)
```

Then when `string` becomes a lower bound of `tout` we flow `string ~> number` and get the error we expect.

This implementation with `TypeDestructorTriggerT` can be used to correctly implement type destructors for all types, however we only use it for open tvars. If we see a concrete type or a resolved tvar then we return a resolved type which is the result of the destructure operation on that resolved type. This is an optimization to prevent evaluating destructors twice when it isn't necessary.

* * *

This diff also updates some library definitions which were inconsistent with their usages. The main library change is to `Promise.all()`. The definition:

```
static all<Elem, T: Iterable<Elem>>(promises: T): Promise<$TupleMap<T, typeof $await>>;
```

Is changed to:

```
static all<T: Iterable<any>>(promises: T): Promise<$TupleMap<T, typeof $await>>;
```

With this change we now handle the upper bounds on the tvars consumed and created by type destructors. This means whereas before `Elem` only got a few, different, lower bounds, now it also gets upper bounds for the same types. The lowers and uppers on `Elem` are incompatible and we start getting errors. (Remember that `$TupleMap` uses the `$Call` type destructor on `Elem` with `typeof $await`.)

We only care that `T` is an iterable. We don’t necessarily care that `Elem`s are compatible. If we assert that `Elem`s are not compatible by factoring `Elem` into a type variable then the following code produces an error:

```
Promise.all(['foo', 42]);
```

Which is undesirable and causes just about all of the error diff.

Reviewed By: avikchaudhuri

Differential Revision: D5827319

fbshipit-source-id: a7c9e196f0e2e20b5b7ee777562665accc3d6a66