Unwrap optional properties in $ObjMap and $ObjMapi
Summary:
Currently, `$ObjMap` has incorrect behavior in the following case:

```
declare var o: $ObjMap<{a: number, b?: number}, <T>(T) => Array<T>>;

(o.a: Array<number>); // OK
(o.b: Array<number> | void); // Should be ok...
(o.b: Array<number | void>); // Should habe an error...
```

When we see an optional property, we treat it as `T | void`. In the example above this means `T | void` is what is passed to our function and therefore is the type wrapped in an array. This behavior is incorrect given the value intuition of `$ObjMap`. If we assume the following implementation (or a similar implementation as `$ObjMap` currently does not respect the-owness of properties):

```
function mapValues(source, mapFn) {
  const result = {};
  for (let p in source) {
    if (Object.hasOwnProperty.call(source, p)) {
      result[p] = mapFn(result[p], p);
    }
  }
  return result;
}
```

Then an optional property in `source` should also be an optional property in `result`. If the property did not exist in `source` there would be no way to set it in `result`.

Reviewed By: samwgoldman

Differential Revision: D5902211

fbshipit-source-id: 8449df45027412af68e976b490cb7569a8aa5266