Use a specific polarity in LookupT for JSX intrinsics
Summary:
The `LookupT` was really bad for perf. The only reason I used a `LookupT` instead of a `GetPropT` is that we needed to add an upper bound to the props type variable. This diff removes the `LookupT` and uses an `AnnotT` to resolve the result of the intrinsics `GetPropT` so that we add an upper bound to the props type variable of a resolved type instead of a type with just a lower bound.

It also looks like there may have been some null prototype termination bugs the Sam found from a TODO I remove in this diff.

Reviewed By: samwgoldman

Differential Revision: D5732538

fbshipit-source-id: ae19ff636da40456ea236e3bb7413e137d887c50