Generalize unwrap_use_ops
Summary:
Before this diff, unwrap_use_ops returned a tuple of reasons and a string msg,
which was useful to pass into typecheck_error, which expects data in that form.

However, it's useful to return top-level errors that do not have two reasons.
Specifically, for FunImplicitReturn errors we want to return a single-line
error message that only points to the ureason, because the lreason location is
arbitrary (it's implicit after all).

In a follow-up, I will add support for FunImplicitReturn unwrapping. I'm
putting this up first to confirm that the change has no unwanted side-effects.

Reviewed By: mroch

Differential Revision: D5919830

fbshipit-source-id: a2e4e36f6f939722cfcc727e7ff9987760730385