Unwrap FunImplicitReturn use ops
Summary:
Before this diff, the errors for an incompatible implicit return were good for
ordinary functions, but bad for async functions and generators.

The issue was that async functions wrapped the `FunImplicitReturn` use op with
`TypeAppCompatibility`. In shorthand, we'd see constraints like this:

```
TypeAppT X, UseT (FunImplicitReturn, TypeAppT Y)
X, UseT (TypeAppCompat FunImplicitReturn, Y
```

If `X` is not a subtype of `Y`, we would hit the `EIncompatibleWithUseOp` error
case, which unwraps use ops. However, the unwrapping code didn't have logic for
`FunImplicitReturn` use ops.

Reviewed By: mroch

Differential Revision: D5919829

fbshipit-source-id: 575cde38502f9a4ee96b71947a6b926cc3fadc44