ssa perf 2
Summary:
`perf` revealed that a bunch of time was being spent in string comparisons, especially in `SMap.find` calls.

As it turns out those calls were due to walking pairs of environments that have the same keysets, in sync, something like what `List.iter2` does for a pair of lists. Even the potentially expensive operation of converting the maps to lists and then using `List.iter2` turns out to be better that the alternative, since it goes from `O(N log N)` to `O(N)`. Presumably GC might now be an issue, because of all those short-lived lists.

Reviewed By: samwgoldman

Differential Revision: D5948909

fbshipit-source-id: 60a9d873abe2e520cd935011494bc6b72b5443b7