Implement $Diff with the new object rest technology
Summary:
This diff (pun intended, again) implements `$Diff` in terms of `$Rest` which I implemented in a previous commit. See that commit for more information about `$Rest`.

`$Diff` should behave mostly the same after this re-implementation. There may be some new errors now that we more strictly compare object properties that are being removed. One common gotcha that I expect we will see a good deal of:

```
$Diff<{}, {b: number}>
```

This is an error\! `void` is incompatible with `number`. To understand why this is an error, let us look at the value semantics:

```
const {b} = {};
(b: number);
```

This would be an error for the same reason. `void` is incompatible with `number`.

This assertion is useful, but it may be confusing to our users at first and could add new errors. The fix is easy:

```
$Diff<{}, {b: number | void}>
```

Reviewed By: samwgoldman

Differential Revision: D5804573

fbshipit-source-id: 3422f752e6a28c4ff8e70e00d91e67714a173115