Make Hh_logger also write to log file for `flow server`
Summary:
When you run `flow start`, the spawned server redirects its
stderr to a file. When you run `flow server`, the server's stderr goes
to stderr. This is a little annoying.

The easiest fix is to just make Hh_logger write to both the log file and
stderr when we're running `flow server`. This is quite a bit easier than
trying to rewrite `tee`.

Reviewed By: samwgoldman

Differential Revision: D5895162

fbshipit-source-id: f3c2a78717bfe1745a08b3509a91f4c4c9c10946