match failure for disjoint unions
Summary:
Let `x: { type: 'A', ... } | { type: 'B', ... }`.

Earlier, when we had

```
if (x.type === 'C') {
  ...
}
```

we'd have `x: empty` in that branch, which was sound but not satisfactory. In
particular, we could miss detecting dead code after refactoring (say before we
had `x: { type: 'C', ... } | { type: 'D', ... }`).

Instead, now we report an early error, saying that 'C' is neither 'A' nor 'B'.

Implementation:

We know that a refinement on a type either passes in the original type or empty,
but where appropriate we could also pass in the *expectation* encoded by the
check, and enforce that the expectation is a subtype of the original type
(because all refinements are supposed to be that way).

In this diff, a failing `(x.type === 'C')` check is considered to encode the
minimum expectation that the type of `x.type` contains `'C'`.

Reviewed By: samwgoldman

Differential Revision: D5971535

fbshipit-source-id: 888d0cd3ce2bc0decfb5a187ae07074e235dd2cf