Harden `flow server` against `flow ide` dying on Linux
Summary:
`flow ide` disappearing should never cause `flow server` to
crash. However, there were a bunch of ways that `flow server` would
crash when `flow ide` died. This diff tries to fix all the cases which I
can repro on Linux.

I found 2 main problems

1. Shutting down an already dead connection crashes, due to a bug in `ServerUtils.shutdown_client`
2. Writing to a closed socket causes an EPIPE error

The first was an easy fix. The second a little trickier, due to the way
that persistent connections are passed around. Once we fail to write or
read to a client, we must never try to write or read to the client
again. Due to the way we stream errors to the client, if we fail to send
even one message, the IDE might get corrupted (like if the EndRecheck
message fails to send).

Reviewed By: samwgoldman

Differential Revision: D5980803

fbshipit-source-id: b468e614b8ed077db21e7f8eedb972532776b2ec