`flow ls path/to/file.js` shouldn't list implicitly ignored files.
Summary:
When you run

  flow ls --root path/to/root file.js

It should list `file.js` if and only if `file.js` is a lib file, a
config file, implicitly included, or explicitly included. Implicitly
ignored or explicitly ignored files are only listed with the `--all`
flag.

However, there was a bug where implicitly ignored files (files outside
of the root directory which are not explicitly included or explicitly
ignored) were being listed. This fixes.

Reviewed By: samwgoldman

Differential Revision: D6042153

fbshipit-source-id: 9056be1701f87b2752c7e4cdf3adacb0bd55de72