`flow server` should survive broken pipe from client
Summary:
I noticed this when testing v0.57.1 but I don't think it's new.
Nevertheless, I wanted to fix it.

Here's what I think happens:

1. A client command connects to the server successfully
2. The client dies before getting a response
3. The server dies when it tries to send the response

I couldn't find a client command for which a failed response should kill
the server. The server's state should be fine even if it can't respond.
So I've caught the broken pipe.

Reviewed By: avikchaudhuri

Differential Revision: D6052181

fbshipit-source-id: 84f429cdd108e238b9ec620e4e07722cd8cbda2e