Allow workers to log to the log
Summary:
Currently `Hh_logger` will log to both stderr and to the log
file when called on the master process. However, when called by a
worker, it only logs to stderr. This diff allows workers to log to the
log as well.

"But Gabe, can't workers race and you end up with interlaced output in
the log?" Yes. It's possible. But that's already possible with stderr.
And as long as our workers aren't logging that heavily, it shouldn't
really be a big problem.

I think the benefit of letting workers log will outweigh the danger

Reviewed By: samwgoldman

Differential Revision: D6048505

fbshipit-source-id: b2f397ed864cf6c3fb43a9785df11baf3635ec71