Print to the log on really slow merges
Summary:
Problem: When flow gets stuck merging, there's no way to know
what it was merging when it got stuck.

We could print which component is getting merged before each merge, but
that would be super verbose, and we don't care about the components
which merge quickly.

We could print after the merge which components took super long, but
that wouldn't help when Flow gets stuck or while Flow is running.

This solution sets a timer that runs after 15 seconds and gets cancelled
when each component is finished being merged. When the timer fires, it
logs which component is taking so long. This way we can see which
components are taking so long and what the server is doing when it gets
stuck!

I initially tested it by logging every 5 seconds, but bumped it up to 15
seconds to avoid noise.

Reviewed By: avikchaudhuri

Differential Revision: D6048544

fbshipit-source-id: 8c56544b8d3d3ed5eee0a9958968a49f2494f5d3