Don't mess up the identifier location for DeclareFunction
Summary: I'm not sure why this was done. It was clearly deliberate, but it makes it so the location of the identifier spans both the identifier and the type annotation. In every other case I've encountered it is the location of the identifier itself. This is weird and inconsistent. Tests pass with this line removed, so it seems like nothing is relying on this behavior.

Reviewed By: gabelevi

Differential Revision: D6054810

fbshipit-source-id: 1a7743cebcfd9db2e80b7f36d92d6e396b04b3e8