Walk choice kit tvars in the type visitor
Summary:
The `id` in a `FullyResolveType` constraint is both an index into the evaluated
map and also a key in the type graph, which points to dependencies/dependents
that are also indexes into the evaluated map (and recursively keys in the type
graph).

The `id` in an `EvalDestructor` constraint points to a tvar. Specifically, it
points to the tvar of an `AnnotT` that appeared in a type destructor.

In addition to adding these cases, I also moved some code from nested match
expressions into private methods. I also reused the existing `type_map` method
in a position where it's body was duplicated.

Reviewed By: avikchaudhuri

Differential Revision: D5888065

fbshipit-source-id: 568f4ed325f6c188fe542b9ac605622e9c395b4b