Rewrite Gc_js as a type visitor
Summary:
I am going to augment the type visitor with logic to track polarities. As noted
in a comment in this file, a polarity-aware type visitor would enable us to do
much more aggressive GC on contexts.

While I don't intend on turning on context GC in production, I do plan to use
GC as a kind of integration test for the polarity logic. If we calculate the
polarities wrong and use the wrong polarities to do GC, we should observe
crashes due to pruned tvars appearing in constraints.

Reviewed By: avikchaudhuri

Differential Revision: D5888069

fbshipit-source-id: 9fb06301ef78ecd1d35b171fbc74713ac72b9378