Make type visitor sensitive to polarities
Summary:
This diff adds a parameter to methods in the type visitor. The base class
implements logic that ensures the value of that parameter reflects the polarity
of that type in that position.

Determining polarities for all types at once is daunting, but happily we can
get away with conservatively estimating the polarities instead. When we don't
know what the polarity should be, either because it's unclear or because I just
haven't gotten around to it yet, I passed `pole_TODO` which is an alias for
`Neutral`.

Reviewed By: avikchaudhuri

Differential Revision: D5888068

fbshipit-source-id: fceb5a4184abd691995002bcc2b16616f76cf7d4