Prune bounds in context GC based on polarities
Summary:
Per the comment that has lived in this file forever, we can prune more during
GC if we track polarities of types that we visit. In particular, if a tvar only
ever appears in a positive position, we can freely collect its upper bounds.
Conversely, if a tvar only appears in a negative position, we can collect it's
lowers.

Reviewed By: avikchaudhuri

Differential Revision: D5888066

fbshipit-source-id: 74a55f3614559d1c00f3e3ec2ad8659bcef07e14