internal types
Summary:
Distinguish types that never appear in signatures. These are presumably produced
and (immediately) consumed in flow_js.ml.

For now, we found only a few, by using a throwaway type visitor on signatures
(not shown). In an ideal world we would have a small subset of types that can
appear in signatures, namely those that are created in
type_annotations.ml. We'll get there someday.

Reviewed By: samwgoldman

Differential Revision: D6050355

fbshipit-source-id: 64ec6dc33712576f43af127f69fa3adea5b002cb