Disable serve_ready_clients
Summary:
`serve_ready_clients` was some code that tried to let the server
respond to some clients in the middle of the recheck. This was always super
dangerous.

With lazy mode, many Flow commands may trigger a recheck of dependencies. Now
that we are lazy by default with regards to node_modules, this is on for
everyone. So if we service something like `flow autocomplete` during a merge,
we might trigger another merge. This somehow causes the Flow server to get
stuck (not sure how).

This diff just disables the behavior. A later diff can actually clean up this
code.

Reviewed By: samwgoldman

Differential Revision: D6064767

fbshipit-source-id: ad6be98e1e93835a07c148f457213ad5e30b071a