Reposition and change the reason description of Object.assign() results
Summary:
It is a little confusing that we point to only the first object in an `Object.assign()` call instead of the entire call. This diff changes our logic so that we point to the full call and not just the first object. See the tests.

```
(Object.assign({}, {a: 42}): {b: number});
```

[try-Flow](https://flow.org/try/#0BQeQRgVgpgxgLgOgIYGcUEsDmA7YBvAXwBoACPJALhIBYAmAgSirzCuwFcBbMKAJ0YDcQA)

Reviewed By: mroch

Differential Revision: D5989613

fbshipit-source-id: 63f30888ef1c0241466c1e5adb64aa7a0c3a57bf