Kill DiffT and re-implement React config checking with new ObjKitT variant
Summary:
Now that `$Diff` is implemented in terms of `$Rest` we can remove the internal `DiffT` type and re-implement the React config checking code. I also took this opportunity to combine what was two object spreads into a single `ObjKitT` operation.

So now we've gone full circle and I've left the object spread code in (mostly) the same state I found it in :)

Reviewed By: samwgoldman

Differential Revision: D5804586

fbshipit-source-id: 9484243b5de1df0c63aaad9aa393ce6c6deb56d3