Global find-refs
Summary: This is a minimum viable product. It only works with ES imports/exports right now, and does not support default exports. It also does not do the equivalent of a get-def before checking to see if the symbol in question has been exported, so it will only find global references if it's called from the file where the symbol is actually exported (as opposed to a file where it is just used). These are all areas for future improvements but this can still be useful in its current form.

Reviewed By: avikchaudhuri

Differential Revision: D5998268

fbshipit-source-id: be535edd3fdeebbf69f602479617133a285ecd82