[hack] Move LSP into utils
Summary:
This is in preparation for sharing the definitions of LSP with Flow.

It feelss weird to put general-purpose stuff like LSP into an "ide" directory, and it feels weird to share an "ide" directory with Flow.

Flow already has machinery set up for sharing files from the utils directory.

Reviewed By: arxanas

Differential Revision: D6008719

fbshipit-source-id: 352d187febfe90288231b7b669606557e91b4758