[hack] Add local configuration option for number of workers and bucket size
Summary:
This adds a config in the local hh.conf that can configure the maximum number of workers.

We also have a flag --max-procs that does the same thing. When given both options, we choose the minimum of the two. The maximum is always at most the number of cores on the machine.

Reviewed By: dabek

Differential Revision: D6104228

fbshipit-source-id: eca1e22602a182f1f861bc060a34f881edbcce9e