Delete remaining serve_ready_clients code
Summary:
This cleans up after {D6064767}

The idea behind serve_ready_clients was that there are some
requests that we can fulfill in a completely safe way without mutating
any state. This has caused trouble in a couple ways over the last month

1. We forgot to blacklist a `flow force-recheck`, which tries to mutate state
2. All the "safe" requests are no longer "safe".

In lazy mode, whenever we run `typecheck_contents`, we may need to pull
in and typecheck any untyped dependencies. And now that we treat
node_modules/ lazily, every mode is lazy mode. So, in theory, we'd need
to blacklist almost every request.

Ultimately, the flow server's was written in a single threaded mindset.
Trying to simulate another thread is possible but it's likely to keep
causing trouble. We absolutely should try and satisfy requests in the
midst of a recheck, but we should build it in a safer way that
guarantees nothing will be mutated.

Reviewed By: avikchaudhuri

Differential Revision: D6066813

fbshipit-source-id: a61c407b39a092ea916a69aced2eed9376c44bc1