Gracefully handle dfind dying
Summary:
Looking at appveyor logs to see why it's so flaky, I'm seeing
that dfind is occasionally dying. I don't have a way to go grab the
dfind logs, so the easiest solution would be to handle the exception,
print the logs, and exit with the correct exit code.

On Windows it gets a `End_of_file` exception from Timeout (translated from
an EPIPE), and on Linux it gets a `Sys_error("Broken pipe")`.

Reviewed By: mroch

Differential Revision: D6054827

fbshipit-source-id: 5e39a0fe8a778dcc6a7cedfd295727c19520cc04