Fix typecheck_contents perf regression with short-circuit
Summary:
`ensure_checked_dependencies` has the very important job of making
sure that all dependencies needed for a merge have already been checked.
However, this is often a no-op. It turns out a no-op call to `typecheck` is
somewhat expensive, so let's short-circuit it if all the dependencies have
already been checked

Reviewed By: mroch

Differential Revision: D6160186

fbshipit-source-id: 51601ddb8694dccab6753ea02931f6e992eb3df1