Prevent flow ide deadlock on OSX and Linux
Summary:
It's possible for both `flow ide` and `flow server` to try and
write large messages to each other and both get blocked. This sometimes
causes the `flow server` to get stuck during merge. It seems quite rare
and difficult to repro.

The easiest fix is to interrupt the `Unix.write` after 5 seconds. This
will cause the client to be marked as broken and later closed. Luckily,
`flow ide` is stateless, so it's not a huge problem when it dies.

Reviewed By: nmote

Differential Revision: D6113948

fbshipit-source-id: b2068e20c368ad957a739b40118e7ef98fddf372