Fix typecheck_contents perf regression with short-circuit
Summary:
`ensure_checked_dependencies` has the very important job of making
sure that all dependencies needed for a merge have already been checked.
However, this is often a no-op. It turns out a no-op call to `typecheck` is
somewhat expensive, so let's short-circuit it if all the dependencies have
already been checked

Reviewed By: mroch

Differential Revision: D6160186

fbshipit-source-id: 031d97c89b799b0d2bd9424d9666ed62890b8c14