[hack] Informant-induced restart can include saved state target
Summary:
When Informant does XDB table lookup for a saved state and finds one,
the server itself doesn't need to do that anymore. This diff passes
the saved state target from the informant to start the new server.

This is part 2 - the prior change consumed the provided saved
state info

Reviewed By: jamesjwu

Differential Revision: D6094120

fbshipit-source-id: abf1f46aadcbb6dfd689f9a82c859a997ac3946d