Replace digest with totally overengineered streaming xxHash
Summary:
Avik did some profiling and saw lots of time spent in sigHash, which calls into
the OCaml Digest module quite a lot.

MD5 is pretty slow, but also the Digest API does not expose a useful
incremental API to initialize a hash, update it a bunch of times, then extract
a digest.

Just using xxHash is better than Digest, due to how hot this code is. However,
we can do even better by avoiding allocations and unnecessary marshalling. This
diff also ensures that we only hash strings and ints.

Because we don't allocate any memory in the update functions, the extern
declaration is annotated with [@noalloc], which makes the call cheaper and
enables the OCaml GC to run during the call.

Reviewed By: avikchaudhuri

Differential Revision: D6145217

fbshipit-source-id: 5ae4ee6c89418317875c8a518391600907428692