Sig hash integers directly instead of caml_output_value_to_block
Summary:
The overhead of OCaml serialization is a fairly large (20 or 32 byte) header.
Since we insist that this value is an integer, we know it should only take 8
bytes.

xxHash has a 32 byte buffer, so it's worthwhile to minimize the size of these
updates in order to do less work.

The way I'm doing this is not portable, but that's fine because the hashes are
local to a single Flow server on a single machine.

Reviewed By: avikchaudhuri

Differential Revision: D6165044

fbshipit-source-id: 6303c3e25c950d26fbbe801298ef740845388187