[hack] Complete switchover to Build_mismatch_ex
Summary:
This diff completes a transition started in Spring 2017 to make the "Build_id_mismatch" carry more information.

While at it, I also augmented the user-facing error report so that when it says "server and client mismatch" to the user, it will now actually tell you which version of client as well as server it has!

Reviewed By: alexchow

Differential Revision: D6080305

fbshipit-source-id: d888fec0a0b66660e782da7c465960eca6cfffdd