Assert that exported class statics are annotated
Summary:
This resolves an oversight in the existing assert_ground visitor, which never
walked class statics. Just like instance properties, we should insist that
static methods and properties are annotated.

Without asking, tvars in statics will get pinned down to the union of lower
bounds. It's very likely that a tvar will get no lower bounds, in which case we
resolve the tvar to `any`, which will miss errors.

Reviewed By: avikchaudhuri

Differential Revision: D6056032

fbshipit-source-id: 11be1130a1a13b99e10fd6c95f3754e98c8ac772