block MatchingPropT ~> BecomeT
Summary:
consider this:

```
// flow
declare function keyMirror<T: {}>(obj: T): $ObjMapi<T, <K>(K) => K>;
type Key = $Keys<typeof Obj>;
const Obj = keyMirror({
  X: null,
  Y: null,
});
function f(value: Key): Key {
  switch (value) {
    case Obj.X: return Obj.Y;
    default: throw "foo";
  }
}
```

in `typeof Obj`, `Obj` is a tvar (call it `T`) because the definite assignment on the next line hasn't happened yet. so the `typeof Obj` flows `T ~> BecomeT(U)` and returns `AnnotT U`.

the switch statement triggers a sentinel property check involving `T`, which generates a `MatchingPropT` and flows it as a lower to `T`. this obviously flows into the `BecomeT`, which then causes an error when `BecomeT` tries to unify the `ObjT` with the `MatchingPropT`.

so, this diff is a hack that blocks the `MatchingPropT` from flowing into a `BecomeT`, because `MatchingPropT` is a refinement of the existing type and the existing type already went through the `BecomeT`, so even if it was safe to do so, it would at the very least be redundant.

Reviewed By: avikchaudhuri

Differential Revision: D6185332

fbshipit-source-id: 05a9e77fe6a5eae735b3b8e16f92cbb6d7d8ed3a