Don't skip AnnotT repositioning during intersection preprocessing repositioning
Summary:
Flow performs some concretization for intersections. For example, when calling
an overloaded function, arguments are concretized first. If one of those
arguments was an AnnotT, we would not concretize it.

The underlying idea here is that IntersectionPreprocessingKit+ConcretizeTypes
constraints should perform the underlying logic appropriate for the lower
bound. We do this for unions (splitting the union) et al., but not for AnnotT.

Reviewed By: mroch

Differential Revision: D6222553

fbshipit-source-id: 794405dc09278880d487d9f80ad22caff8a30282