fix recursion limit error for polymorphic instanceof
Summary:
Doing `x instanceof` on polymorphic classes executes the check on a fresh
instance, which can cause non-termination because when the check "fails" we
flow the fresh instance as the refinement of x: if now we do some operation on
this x and feed the value back into the check, we are in trouble.

There are many possible solutions to this problem. This diff picks one that was
also picked for React.Props.InstanceOf.

Reviewed By: samwgoldman

Differential Revision: D6231177

fbshipit-source-id: ad693ca8cc7ab3ddf52196a944585ef5333ab59e