Avoid mem+add in FlowConstraint cache
Summary:
As of OCaml 4.03, add will return a referentially equal set if you try to add a
value that is already in the set. We can use this to avoid walking the set
twice, doing potentially expensive comparisons along the way.

I.e., this improves a constant factor, but the constant can be pretty big. The
cost corresponds to the size of the use_t being compared.

Reviewed By: mroch

Differential Revision: D6236462

fbshipit-source-id: 73bd1793b978e126c0b51145f4df32b35196874b