SocketHandshake
Summary:
When we connect over the socket, a handshake is performed. This
is currently very simple.

1. The client sends the build id
2. The server either responds "Connection_ok" or "Build_id_mismatch"

I'm making this a little more sophisticated.

1. The client sends the build id and what kind of connection it wants to establish
2. The server responds either "Connection_ok" or "Build_id_mismatch"

In the future, I want to change the build id mismatch behavior. Instead
of the server dying, I want it to tell the client which Flow binary to
use. Then the client can exec that binary and try again.

Reviewed By: samwgoldman

Differential Revision: D6247662

fbshipit-source-id: 48bdb93468eb3c069f3376af3508cf62c97c6655